+++ Week-5/Excerise-5.4/calculator.test.js
@@ -1,0 +1,13 @@
+const mathOperations = require("./calculator");
+
+describe("calculator tests", () => {
+  test("Adding 1+2 should return 3", () => {
+    expect(mathOperations.sum(1, 2)).toBe(3);
+  });
+  test("Diff 100 -50 should return 50", () => {
+    expect(mathOperations.diff(100, 50)).toBe(50);
+  });
+  test("Product 9*2 should return 18", () => {
+    expect(mathOperations.product(9, 2)).toBe(18);
+  });
+});
