--- Week-5/Excerise-5.4/index.js
@@ -1,12 +1,0 @@
-const mathOperations = {
-  sum: function (a, b) {
-    return a + b;
-  },
-  diff: function (a, b) {
-    return a - b;
-  },
-  product: function (a, b) {
-    return a * b;
-  },
-};
-module.exports = mathOperations;
