+++ Week-5/Excerise-5.4/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "Excerise-5.4",
+  "name": "5.4",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
